Refactor inference into runtime workflow modules
