+++ slide2vec/runtime/__init__.py
@@ -1,0 +1,2 @@
+"""Internal runtime helpers for slide2vec inference orchestration."""
+
